add virtual receipt page and refactoring handlers

+++ cmd/web/render.go
@@ -32,7 +32,7 @@ var functions = template.FuncMap{
 }
 
 func formatCurrency(n int) string {
-	f := float64(n / 100)
+	f := float32(n) / float32(100)
 	return fmt.Sprintf("$%.2f", f)
 }
